test: migrate zero-uploads-prepare to bdd
